+++ src/search.h
@@ -26,7 +26,7 @@ ag_stats stats;
 size_t skip_lookup[256];
 
 void search_buf(const pcre *re, const pcre_extra *re_extra,
-                const char *buf, const int buf_len,
+                const char *buf, const long buf_len,
                 const char *dir_full_path);
 void search_stdin(const pcre *re, const pcre_extra *re_extra);
 void search_file(const pcre *re, const pcre_extra *re_extra, const char *file_full_path);
