+++ src/util.c
@@ -70,7 +70,7 @@ char* boyer_moore_strncasestr(const char *s, const char *find, size_t s_len, siz
 }
 
 /* TODO: haven't slept in 48 hours. need to look this code over when I'm less stupid */
-int invert_matches(match matches[], int matches_len, const int buf_len) {
+int invert_matches(match matches[], size_t matches_len, const size_t buf_len) {
     int i;
 
     /* this will totally screw-up if a match starts at the very beginning or end of a file */
@@ -84,11 +84,11 @@ int invert_matches(match matches[], int matches_len, const int buf_len) {
     return(matches_len + 1);
 }
 
-int is_binary(const void* buf, const int buf_len) {
-    int suspicious_bytes = 0;
-    int total_bytes = buf_len > 1024 ? 1024 : buf_len;
+int is_binary(const void* buf, const size_t buf_len) {
+    size_t suspicious_bytes = 0;
+    size_t total_bytes = buf_len > 1024 ? 1024 : buf_len;
     const unsigned char *buf_c = buf;
-    int i;
+    unsigned int i;
 
     if (buf_len == 0) {
         return(0);
